Add files via upload
First entry. These are the machine/hardware dependent files. All in UniFLEX format. To build: first in ../kernel    update_all,
afterwards in this directory  update_all. Beware that change in the ../include files are not noticed. Safest way is to kill all .r files first
and then rebuild kernel and machine files. uniflex-IDE is the resulting kernel. Copy it to the correct place under the correct name!